completed player pref manager
completed player pref manager to include level completion durations, total stars per level and which type of star per level, with functions to add new stars, check for duration times and save the faster of old and new times and returns a bool for true if faster, but false in other cases. Tested and tweaked. Looks good to go.