minor changes to dog script
Made a couple minor changes to dog script so it acting properly. There was a bug where chasing wanted to call a bark hash for "Barking" when the animator was using "Bark" so the bark was not getting properly trigger unless inside the cats body. Made a change so chase would cross lanes to directly the cat z position. tested this in level 15